refactor: move runs artifact cleanup into service
